third_party/WebKit/LayoutTests/platform/win/external/wpt/html/browsers/history/the-location-interface/location-protocol-setter-non-broken-expected.txt

CONSOLE ERROR: Not allowed to load local resource: blank.html
CONSOLE WARNING: Subresource requests using legacy protocols (like `ftp:`) are deprecated, and will be blocked in M59, around June 2017. Please deliver web-accessible resources over modern protocols like HTTPS. See https://www.chromestatus.com/feature/5709390967472128 for more details.
CONSOLE ERROR: Not allowed to load local resource: script%3E
CONSOLE ERROR: Not allowed to load local resource: script%3E
This is a testharness.js-based test.
PASS Set HTTP URL frame location.protocol to x 
PASS Set data URL frame location.protocol to x 
FAIL Set HTTP URL frame location.protocol to data Blocked a frame with origin "http://web-platform.test:8001" from accessing a cross-origin frame.
FAIL Set data URL frame location.protocol to data Failed to execute 'postMessage' on 'Window': Port at index 0 is already neutered.
FAIL Set HTTP URL frame location.protocol to file Failed to set the 'protocol' property on 'Location': 'file://web-platform.test:8001/common/blank.html' is not a valid URL.
PASS Set data URL frame location.protocol to file 
PASS Set HTTP URL frame location.protocol to ftp 
PASS Set data URL frame location.protocol to ftp 
PASS Set HTTP URL frame location.protocol to gopher 
PASS Set data URL frame location.protocol to gopher 
PASS Set HTTP URL frame location.protocol to http+x 
PASS Set data URL frame location.protocol to http+x 
Harness: the test ran to completion.
